requirements.txt

langchain
langchain-community
langchain-ollama
beautifulsoup4
requests
scikit-learn
chromadb
streamlit
pillow
pytesseract
